+++ prefilter/teddy_fat_fallback.go
@@ -2,12 +2,15 @@
 
 package prefilter
 
+// hasAVX2 is always false on non-x86-64 platforms.
+var hasAVX2 = false
+
 // findSIMD falls back to scalar search on non-x86-64 platforms.
-//
-// Fat Teddy requires AVX2 which is only available on x86-64.
-// On ARM, RISC-V, etc., use scalar implementation.
-//
-// Future: ARM NEON implementation possible (Go 1.26+ SIMD support)
 func (t *FatTeddy) findSIMD(haystack []byte) (pos int, bucketMask uint16) {
 	return t.findScalarCandidate(haystack)
+}
+
+// fatTeddyAVX2_2Batch is unavailable on non-x86-64 platforms.
+func fatTeddyAVX2_2Batch(_ *fatTeddyMasks, _ []byte, _ []uint64) int {
+	return 0
 }